Dreamhack/secure and simple/src/app/requirements.txt
flask
pymongo
pycryptodome
sympy